05-findify/pyproject.toml
[project]
name = "findify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.0",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.7",
    "langchain-pinecone>=0.2.8",
    "mysql-connector-python>=9.3.0",
    "pinecone>=7.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
]

[tool.uv.workspace]
members = [
    "backend",
    "frontend",
]
